requirements.txt
streamlit>=1.35.0
google-genai>=1.0.0
openai>=1.30.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
pdfplumber>=0.11.0
pypdf>=4.2.0
numpy>=1.26.0
